LayoutTests/platform/glib/http/tests/websocket/tests/hybi/handshake-fail-by-null-char-in-status-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/handshake-fail-by-null-char-in-status' failed: The server did not accept the WebSocket handshake.
Connection should fail immediately, rather than succeeding or staying in limbo until timeout, if a null byte is received in the status header.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS timedOut is false
PASS connected is false
PASS wsOrigin is undefined.
PASS successfullyParsed is true

TEST COMPLETE
